% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidarTools.R
\name{getPAI}
\alias{getPAI}
\title{Single return lidar PAI function}
\usage{
getPAI(z, A, zi, ai, lad = "spherical", verbose = FALSE)
}
\arguments{
\item{(z)}{1-D vector of vertical coordinates}

\item{(A)}{1-D vector of scan angles}

\item{(zi)}{1-D vector of discrete z-axis at which leaf area density will be computed
vectors is expected to have a fixed step size (dzi).}

\item{(lad)}{leaf inclination model ('planophile','erectophile','spherical', or user defined)}

\item{(ai)}{scan angle (optional for re-computing leaf area density with different leaf inclination)}

\item{(verbose)}{logical: sends signs of life?}
}
\description{
Single return lidar PAI function
}

