% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidarTools.R
\name{Gfunction}
\alias{Gfunction}
\title{Compute Ross G function}
\usage{
Gfunction(lad = "spherical", ze, zi, par)
}
\arguments{
\item{(ze)}{Incoming angle in degrees}

\item{(lad)}{character: leaf angle distribution (uses radians)}

\item{(zi)}{vector of vertical coordinates (height)}

\item{(par)}{= optional parameters to fit the beta}
}
\description{
Compute Ross G function
}

