% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indexer}
\alias{indexer}
\title{Create a varying index}
\usage{
indexer(i, end)
}
\arguments{
\item{i}{Where the index should start}

\item{end}{Where the index should end (essentially length)}
}
\value{
A vector of the index
}
\description{
This is a simple utility to create a fixed length index.
}
\examples{
indexer(1, 10)
indexer(2, 10)
indexer(3, 5)
}
\keyword{utils}

