% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidarTools.R
\name{indexFromShape}
\alias{indexFromShape}
\title{Extract index of point cloud that are within a shapefile}
\usage{
indexFromShape(xyz, qtr, class = "class", shpfile)
}
\arguments{
\item{(xyz)}{matrix point cloud with x,y, and z coordinates}

\item{(qtr)}{quadtree of xyz}

\item{(shpfile)}{a SpatialPolygonsDataFrame to subset the point cloud
only the first polygon is used if multiple polygons are
present}
}
\description{
Extract index of point cloud that are within a shapefile
}

