% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lm_pai_cvboot_fit}
\alias{lm_pai_cvboot_fit}
\title{Calculate Personalized Advantage Index (PAI)}
\usage{
lm_pai_cvboot_fit(formula, DV, TreatVar, dat, boot = TRUE, k, holdouts,
  yhigherisbetter = TRUE)
}
\arguments{
\item{formula}{An R formula, but without dependent variable}

\item{DV}{a character string indicating the name of the dependent variable}

\item{TreatVar}{a character string indicating the name of the treatment variable
(note that values must be coded as 0/1 for the treatment variable.}

\item{dat}{A dataset to use}

\item{boot}{A logical value whether to bootstrap or not}

\item{k}{The number of bootstrap resamples to take}

\item{holdouts}{The holdouts to be used.  If a vector from 1 to the number of
cases available in the dataset, it is equivalent to leave-one-out.
Must be constructed manually.}

\item{yhigherisbetter}{A logical value whether higher scores indicate better
treatment response on the dependent variable}
}
\value{
A three dimensional array with cases on the first dimension,
  multiple estimates calculated on the second dimension, and bootstrap resamples
  on the third dimension.
}
\description{
Calculates the PAI for each observation in the dataset using simple
linear regression, but with cross-validation and optionally bootstrapping
in order to construct confidence intervals.
}
\examples{
# builtin dataset with 32 cases
dat <- mtcars
dat$cyl <- factor(dat$cyl)
formula <- ~ cyl + am * (mpg + hp + drat + wt)
DV <- "disp"
TreatVar <- "am"
k <- 50 # use a few just for speed
holdouts <- 1:32 # for leave-one-out

# for 10 fold cross-validation
holdouts <- by(sample(1:32), cut(1:32, ceiling(seq(from = 0, to = 32, length.out = 11))), function(x) x)
# remove names
names(holdouts) <- NULL

m <- lm_pai_cvboot_fit(formula, DV, TreatVar, dat, boot = TRUE, k = k, holdouts = holdouts)

# see result from first bootstrap resample
m[, , 1]
}

