% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.mipai}
\alias{predict.mipai}
\title{Calculate Individual PAIs}
\usage{
\method{predict}{mipai}(object, plot = TRUE, cores = 1L, ...)
}
\arguments{
\item{object}{an mipai or mibootpai class object returned from
\code{mi_lm_pai_cvboot}.}

\item{plot}{logical whether to plot a dotplot.}

\item{cores}{An integer, the number of cores to be used for a local cluster
when calculating CIs on bootstrapped models.}

\item{dots}{Additional arguments, not currently used.
a label for the results.}
}
\value{
An invisible list of two elements, the \dQuote{PAI} scores and
  the \dQuote{Graph}, which is returned even if not plotted.
}
\description{
Calculate Individual PAIs
}
\examples{
\dontrun{
# builtin dataset with 32 cases
dat <- mtcars
dat$cyl <- factor(dat$cyl)

m <- mi_lm_pai_cvboot(~ cyl + am * (mpg + hp + drat + wt),
  "disp", "am", list(dat, dat),
  nboot = 50, holdouts = "10", cores = 2)

predict(m, cores = 2L)
}
}

