% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resid.mipai}
\alias{resid.mipai}
\title{Calculate the Residuals for Predictions}
\usage{
\method{resid}{mipai}(object, type = c("MeanAbs", "MedianAbs", "RootMeanSq",
  "RootMedianSq"), summarize = TRUE, boot = TRUE, ...)
}
\arguments{
\item{object}{an mipai or mibootpai class object returned from
\code{mi_lm_pai_cvboot}}

\item{type}{A character string indicating the type of residuals to be calculated.
Currents mean or median of the absolute or root squared values.  Only applies
when summarize is \code{FALSE}.}

\item{summarize}{Logical whether or not to summarize values for individuals
across imputations or leave as a matrix of residuals.  Defaults to \code{TRUE}.}

\item{boot}{Not currently used.}

\item{dots}{Additional arguments to pass down.  Not currently used.}
}
\value{
A vector of residuals for each person or matrix across the
  multiple imputations.
}
\description{
Calculate the Residuals for Predictions
}
\examples{
\dontrun{
# builtin dataset with 32 cases
dat <- mtcars
dat$cyl <- factor(dat$cyl)

m <- mi_lm_pai_cvboot(~ cyl + am * (mpg + hp + drat + wt),
  "disp", "am", list(dat, dat),
  nboot = 50, holdouts = "10", cores = 2)

resid(m, "RootMeanSq")
}
}

