% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidarTools.R
\name{shapesXYZ}
\alias{shapesXYZ}
\title{Get point cloud from a shapefile
 lidar subset returned as SpatialPointsDataFrame}
\usage{
shapesXYZ(xyz, qtr, shpfile, return.index = FALSE)
}
\arguments{
\item{(xyz)}{matrix point cloud with x,y, and z coordinates}

\item{(qtr)}{quadtree of xyz}

\item{(shpfile)}{a SpatialPolygonsDataFrame to subset the point cloud
only the first polygon is used if multiple polygons are}

\item{(return.index)}{logical, return row index for shapefile
from xyz
present}
}
\description{
Get point cloud from a shapefile
 lidar subset returned as SpatialPointsDataFrame
}

