% Generated by roxygen2 (4.0.1): do not edit by hand
\name{statsummary}
\alias{statsummary}
\title{Summary Statistics in a Named Vector}
\usage{
statsummary(x, types, na.rm = TRUE)
}
\arguments{
\item{x}{A data vector}

\item{types}{The types of summaries to be computed.}

\item{na.rm}{Logical passed on to stat functions.}
}
\value{
a named vector
}
\description{
Simple utility wrapper to make it easy to get a variety of summary
statistics in a named vector, in whatever order desired.
}
\examples{
statsummary(1:10, c("Mean", "Min", "Max", "IQR", "Kurtosis"))
}
\keyword{utils}

