% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mi_lm_pai_cvboot}
\alias{mi_lm_pai_cvboot}
\title{Calculate PAI on multiply imputed data in parallel}
\usage{
mi_lm_pai_cvboot(formula, DV, TreatVar, imputedData, boot = TRUE, k = 500,
  holdouts, yhigherisbetter = TRUE, seed = 1234, cl, cores)
}
\arguments{
\item{formula}{An R formula, but without dependent variable}

\item{DV}{a character string indicating the name of the dependent variable}

\item{TreatVar}{a character string indicating the name of the treatment variable
(note that values must be coded as 0/1 for the treatment variable.}

\item{imputedData}{An imputed dataset (list of datasets) to use}

\item{boot}{A logical value whether to bootstrap or not}

\item{k}{The number of bootstrap resamples to take}

\item{holdouts}{The holdouts to be used.  If a vector from 1 to the number of
cases available in the dataset, it is equivalent to leave-one-out.
Must be constructed manually.}

\item{yhigherisbetter}{A logical value whether higher scores indicate better
treatment response on the dependent variable}

\item{seed}{A seed to make the bootstrap reproducible}

\item{cl}{An existing cluster to use (optional)}

\item{cores}{The number of cores to use when creating a cluster, if an existing
cluster is not passed (optional).  If left blank and no cluster passed,
defaults to the number of cores available.}
}
\value{
A three dimensional array with cases on the first dimension,
  multiple estimates calculated on the second dimension, and bootstrap resamples
  on the third dimension.
}
\description{
This is a simple function that wraps \code{lm_pai_cvboot}.
}
\examples{
# make me!!!
}

