% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.mipai}
\alias{summary.mipai}
\title{Calculate Summary Statistics for PAI Model}
\usage{
\method{summary}{mipai}(object, plot = TRUE, boot = TRUE, ...)
}
\arguments{
\item{object}{an mipai or mibootpai class object returned from
\code{mi_lm_pai_cvboot}.}

\item{plot}{logical whether to plot a histogram of the PAI scores.}

\item{dots}{Additional arguments, not currently used.
a label for the results.}

\item{boot}{logical whether to use bootstrapped results. Defaults to
\code{TRUE} if the object includes bootstrapped results.}
}
\value{
A list of data frames with an element of summary statistics
  for the PAI (PAI) and observed results for patients assigned to
  their optimal and nonoptimal treatment (ObservedOutcome) and
  residuals (Residuals), and for the histogram (Graph).
}
\description{
Calculate Summary Statistics for PAI Model
}
\examples{
\dontrun{
# builtin dataset with 32 cases
dat <- mtcars
dat$cyl <- factor(dat$cyl)

m <- mi_lm_pai_cvboot(~ cyl + am * (mpg + hp + drat + wt),
  "disp", "am", list(dat, dat),
  nboot = 50, holdouts = "10", cores = 2)

summary(m)
}
}

